go internal/lexer: switch Location to return ast node
This seems a little sketchy to me (the lexer importing the ast),
but it seems like it would make the parser code simpler. Perhaps
all of the location-aware bits should use the ast?

Or we hoist Location() into the parser and continue to expose
simple ints, but that seems annoying.