internal/*: support whitespace sensitivity for parsing selectors
It seems like most css doesn't care about whitespace, but some does.

For instance, "div.test" and "div .test" are very different selectors. To
keep this disambiguated, we add an extra mode to the lexer that returns
whitespace tokens instead of swallowing them. In this commit,
the selector parser uses this mode.

This change lets us mostly keep parsing code the same and not have
to deal with whitesapce everywhere while correctly parsing selectors.