internal/transformer: add start of custom properties transform
The transforms code is a bit different than the printer since
you want to be able to replace arbitrary nodes. It's hard to do
that without accidentally getting into reflect code, so instead
each interesting case needs to be handled manually.

Instead, it seems like we want functions to operate on fields for
of each node, instead of functions that operate directly on nodes. This
lets us make decisions from the parents perspective, like cleaning
up unused :root in variable replacement.

Note that this commit isn't complete - we don't properly implement
fallbacks or fallbacks with multiple values.