internal/*: change custom-media to only take one media query
From https://github.com/w3c/csswg-drafts/issues/5458#issuecomment-678358973,
it seems that it's unintended that @custom-media allows multiple
media queries. Instead, we take only one.

I added a test for the case where multiple are returned, but the
error message is unsatisfying - it complains about wanting
a selector (because custom media at rule parsing is done and
we want to move on to parsing a qualified rule), but it
isn't helpful for someone who is confused about the at rule.