internal/{logging,transformer}: add reporter api and use in transformer
There are a few weird things in this commit.

One is that the transformer and printer APIs are diverging. Transformer uses optional
function arguments alongside a required source argument, where as printer puts source
as a single argument. Probably, we normalize later.

Also, the transformer doesn't have access to spans (start and end) of text
for reporting location errors, since they've been discarded by that point. Maybe
we should change ast.Loc to ast.Span and add a length/end so that we can retain it.

I think this also means that we probably want to retain ast.Loc on all
nodes so that error reporting works in transforming too.

As an alternative, perhaps we could move transform passes into lexer/parser stage
(where location information is still around), but I think it doesn't
fully solve the problem because some transforms (variables, @import) have to be
done after the first parsing phase.