api: use Reporter for api output
This commit changes the api to use reporter for parse and transform. A subsequent
commit will do printer separately after changing that API to return errors.

Parser/printer steps can rely on panics because they are all-or-nothing phases. Transformer
can emit errors and warnings and continue, so we pass the reporter down into it.