interal/transformer: switch from function options to options struct
I think this is a bit easier for not having to write a bunch of boilerplate
for all the functions we're going to want to write. It also makes the transformer
api look like the printer api.

Also move api transform options into different package so that they
can be used in both api and transformer.