transforms: add math reduction transform
This commit has a couple issues. One is that evaluteMathExpression
is really hard to read. Working with the types in go is a big pain.

A couple others are captured in the tests. We don't chop off precision
in some cases. We should probably go back and make all precisions
Cut off at the 3/4/5th decimal.

Another issue is that because we cannot reduce all expressions,
we cannot change 22% + 5rem + 2rem into 22% + 7rem, as the evaluator
does not know how to work with (mathexpr + 2rem) at the outer level of
the AST.