internal/*: collapse StartSpan into TokenSpan
I realized that we might as well always use TokenSpan, with
some kinds of AST rewriting the end. This change makes
the parser somewhat more ambiguous (as it's not immediately clear
if the current AST node needs more than the current token) but it seems worthwhile
to keep the API more bare.