internal/parser: add complete an+b support
This code is a big mess because it's hard to deconstruct ambiguities
in css lexer tokens that are not taken into consideration for the
an+b microsyntax. This will have to do for now. We might consider
instead writing a separate parser (not on the current lexer) for an+b
or just retaining the string content without parsing an+b nodes into
a semantically useful node.